Move console package into sleeper.clients
