+++ java/clients/src/main/java/sleeper/clients/admin/AdminCommonPrompts.java
@@ -15,11 +15,11 @@
  */
 package sleeper.clients.admin;
 
+import sleeper.clients.console.ConsoleInput;
+import sleeper.clients.console.ConsoleOutput;
+import sleeper.clients.console.menu.ConsoleChoice;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
-import sleeper.console.ConsoleInput;
-import sleeper.console.ConsoleOutput;
-import sleeper.console.menu.ConsoleChoice;
 
 import java.util.Optional;
 import java.util.function.Supplier;
