+++ java/clients/src/main/java/sleeper/clients/admin/AdminMainScreen.java
@@ -16,12 +16,12 @@
 package sleeper.clients.admin;
 
 import sleeper.clients.AdminClient;
-import sleeper.console.ConsoleInput;
-import sleeper.console.ConsoleOutput;
-import sleeper.console.UserExitedException;
-import sleeper.console.menu.ChooseOne;
-import sleeper.console.menu.Chosen;
-import sleeper.console.menu.MenuOption;
+import sleeper.clients.console.ConsoleInput;
+import sleeper.clients.console.ConsoleOutput;
+import sleeper.clients.console.UserExitedException;
+import sleeper.clients.console.menu.ChooseOne;
+import sleeper.clients.console.menu.Chosen;
+import sleeper.clients.console.menu.MenuOption;
 
 import java.util.Arrays;
 import java.util.List;
