+++ java/clients/src/main/java/sleeper/clients/admin/PropertiesDiff.java
@@ -16,10 +16,10 @@
 
 package sleeper.clients.admin;
 
+import sleeper.clients.console.ConsoleOutput;
 import sleeper.configuration.properties.SleeperProperties;
 import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.SleeperPropertyIndex;
-import sleeper.console.ConsoleOutput;
 
 import java.util.ArrayList;
 import java.util.Collections;
