+++ java/clients/src/main/java/sleeper/clients/admin/PropertyDiff.java
@@ -16,9 +16,9 @@
 
 package sleeper.clients.admin;
 
+import sleeper.clients.console.ConsoleOutput;
 import sleeper.configuration.properties.SleeperProperty;
 import sleeper.configuration.properties.SleeperPropertyIndex;
-import sleeper.console.ConsoleOutput;
 
 import java.util.Map;
 import java.util.Objects;
