+++ java/clients/src/main/java/sleeper/clients/admin/TableNamesReport.java
@@ -15,8 +15,8 @@
  */
 package sleeper.clients.admin;
 
-import sleeper.console.ConsoleInput;
-import sleeper.console.ConsoleOutput;
+import sleeper.clients.console.ConsoleInput;
+import sleeper.clients.console.ConsoleOutput;
 
 import java.util.List;
