+++ java/clients/src/main/java/sleeper/clients/admin/TableSelectHelper.java
@@ -16,14 +16,14 @@
 
 package sleeper.clients.admin;
 
+import sleeper.clients.console.ConsoleInput;
+import sleeper.clients.console.ConsoleOutput;
+import sleeper.clients.console.UserExitedException;
+import sleeper.clients.console.menu.ChooseOne;
+import sleeper.clients.console.menu.Chosen;
+import sleeper.clients.console.menu.ConsoleChoice;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
-import sleeper.console.ConsoleInput;
-import sleeper.console.ConsoleOutput;
-import sleeper.console.UserExitedException;
-import sleeper.console.menu.ChooseOne;
-import sleeper.console.menu.Chosen;
-import sleeper.console.menu.ConsoleChoice;
 
 import java.util.Optional;
