+++ java/clients/src/main/java/sleeper/clients/console/menu/MenuOption.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.console.menu;
+package sleeper.clients.console.menu;
 
-import sleeper.console.UserExitedException;
+import sleeper.clients.console.UserExitedException;
 
 public class MenuOption implements ConsoleChoice {
