+++ java/clients/src/main/java/sleeper/clients/status/report/CompactionJobStatusReport.java
@@ -21,6 +21,7 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 
+import sleeper.clients.console.ConsoleInput;
 import sleeper.clients.status.report.compaction.job.CompactionJobStatusReporter;
 import sleeper.clients.status.report.compaction.job.JsonCompactionJobStatusReporter;
 import sleeper.clients.status.report.compaction.job.StandardCompactionJobStatusReporter;
@@ -30,7 +31,6 @@ import sleeper.clients.util.ClientUtils;
 import sleeper.compaction.job.CompactionJobStatusStore;
 import sleeper.compaction.status.store.job.CompactionJobStatusStoreFactory;
 import sleeper.configuration.properties.InstanceProperties;
-import sleeper.console.ConsoleInput;
 
 import java.io.IOException;
 import java.time.Clock;
