+++ java/clients/src/test/java/sleeper/clients/admin/TableNamesReportTest.java
@@ -26,7 +26,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.MAIN_SCREEN;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.PROMPT_RETURN_TO_MAIN;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.TABLE_NAMES_REPORT_OPTION;
-import static sleeper.console.ConsoleOutput.CLEAR_CONSOLE;
+import static sleeper.clients.console.ConsoleOutput.CLEAR_CONSOLE;
 import static sleeper.console.TestConsoleInput.CONFIRM_PROMPT;
 
 class TableNamesReportTest extends AdminClientMockStoreBase {
